# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Plastic Waste Management API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a citizen or role-scoped user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
...
components:
...
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string }
        password: { type: string, minLength: 8 }
        role:
          type: string
          enum: [citizen, centre, driver]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string, minLength: 1 }
    AuthSession:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/CurrentUser"
    CurrentUser:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
...