# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Plastic Waste Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CurrentUser } from './currentUser';

export interface AuthSession {
  token: string;
  user: CurrentUser;
}