# lib/api-zod/src/generated/types/currentUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Plastic Waste Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CurrentUser {
  id: string;
  name: string;
  email: string;
  role: string;
}