# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Plastic Waste Management API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputRole } from './registerInputRole';

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
  role: RegisterInputRole;
}